moving the GWAS service into one folder
