+++ codecept.conf.js
@@ -84,7 +84,7 @@ exports.config = {
 
     // Pages
     home: './services/portal/home/homeService.js',
-    gwas: './services/portal/GWASUIApp/GWASUIAppService.js',
+    gwas: './services/portal/GWAS/GWASService.js',
     indexing: './services/portal/indexing/indexingService.js',
     login: './services/portal/login/loginService.js',
     explorer: './services/portal/explorer/explorerService.js',
