+++ services/portal/GWAS/GWASService.js
@@ -1,0 +1,11 @@
+const GWASTasks = require('./GWASTasks.js');
+const GWASProps = require('./GWASProps.js');
+
+/**
+ * GWASUIApp Service
+ */
+module.exports = {
+    props: GWASProps,
+
+    do: GWASTasks,    
+};
