--- services/portal/GWASUIApp/GWASUIAppService.js
@@ -1,11 +1,0 @@
-const GWASUIAppTasks = require('./GWASUIAppTasks.js');
-const GWASUIAppProps = require('./GWASUIAppProps.js');
-
-/**
- * GWASUIApp Service
- */
-module.exports = {
-    props: GWASUIAppProps,
-
-    do: GWASUIAppTasks,    
-};
